# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LifePath 2.0 API — The App That Fixes What Life Breaks
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found

    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/onboard:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete onboarding flow and create/update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "201":
          description: Onboarding complete, user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResult"

  # ─── LIFE SCORE ───────────────────────────────────────────────────────────
...
  /urges:
    get:
      operationId: getUrges
      tags: [urges]
      summary: Get urge log history
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Urge logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UrgeLog"
    post:
      operationId: logUrge
      tags: [urges]
      summary: Log a new urge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UrgeInput"
      responses:
        "201":
          description: Urge logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UrgeLog"
...
  /streaks/relapse:
    post:
      operationId: logRelapse
      tags: [streaks]
      summary: Log a relapse and start recovery protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelapseInput"
      responses:
        "200":
          description: Relapse logged, recovery started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Streak"

  # ─── BODY DOUBLING ROOMS ──────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, onboardingComplete, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        biggestStruggle:
          type: ["string", "null"]
        struggleDuration:
          type: ["string", "null"]
        triedBefore:
          type: ["string", "null"]
        changeStatement:
          type: ["string", "null"]
        activeModules:
          type: array
          items:
            type: string
        addictionType:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        activeModules:
          type: array
          items:
            type: string
        addictionType:
          type: string
    OnboardingInput:
      type: object
      required: [name, biggestStruggle, struggleDuration, triedBefore, changeStatement]
      properties:
        name:
          type: string
        biggestStruggle:
          type: string
          enum: [procrastination, compulsive_habits, cant_focus, all_of_the_above]
        struggleDuration:
          type: string
          enum: [weeks, months, years]
        triedBefore:
          type: string
          enum: [yes_nothing_worked, yes_briefly, no_first_time]
        changeStatement:
          type: string
          maxLength: 120
        addictionType:
          type: string
          enum: [pornography, social_media, gambling, substances, compulsive_eating]
    OnboardingResult:
      type: object
      required: [user, initialLifeScore]
      properties:
        user:
          $ref: "#/components/schemas/User"
        initialLifeScore:
          type: integer
          minimum: 20
          maximum: 55

    # ── Life Score ──────────────────────────────────────────────────────────
...
    UrgeLog:
      type: object
      required: [id, haltState, intensity, outcome, createdAt]
      properties:
        id:
          type: integer
        haltState:
          type: array
          items:
            type: string
            enum: [hungry, angry, lonely, tired]
        intensity:
          type: integer
          minimum: 1
          maximum: 10
        outcome:
          type: string
          enum: [resisted, surfed, redirected, relapsed]
        redirectActivity:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    UrgeInput:
      type: object
      required: [haltState, intensity]
      properties:
        haltState:
          type: array
          items:
            type: string
            enum: [hungry, angry, lonely, tired]
        intensity:
          type: integer
          minimum: 1
          maximum: 10
        outcome:
          type: string
          enum: [resisted, surfed, redirected, relapsed]
        redirectActivity:
          type: string
        notes:
          type: string
...
    Streak:
      type: object
      required: [currentDays, longestDays, addictionType, startedAt, milestones]
      properties:
        currentDays:
          type: integer
        longestDays:
          type: integer
        addictionType:
          type: string
        startedAt:
          type: string
        nextMilestone:
          $ref: "#/components/schemas/Milestone"
        milestones:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        dopamineRecoveryPercent:
          type: number
          minimum: 0
          maximum: 100
        neuroBrainFact:
          type: ["string", "null"]
    Milestone:
      type: object
      required: [day, label, scienceFact, achieved]
      properties:
        day:
          type: integer
        label:
          type: string
        scienceFact:
          type: string
        achieved:
          type: boolean
        achievedAt:
          type: ["string", "null"]
    RelapseInput:
      type: object
      required: [triggeredBy]
      properties:
        triggeredBy:
          type: array
          items:
            type: string
        notes:
          type: string

    # ── Rooms ───────────────────────────────────────────────────────────────
...