# lib/api-zod/src/generated/types/relapseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifePath 2.0 API — The App That Fixes What Life Breaks
 * OpenAPI spec version: 0.1.0
 */

export interface RelapseInput {
  triggeredBy: string[];
  notes?: string;
}