# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LifePath 2.0 API — The App That Fixes What Life Breaks
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  /** @nullable */
  biggestStruggle?: string | null;
  /** @nullable */
  struggleDuration?: string | null;
  /** @nullable */
  triedBefore?: string | null;
  /** @nullable */
  changeStatement?: string | null;
  activeModules?: string[];
  /** @nullable */
  addictionType?: string | null;
  onboardingComplete: boolean;
  createdAt: string;
}